[unreal]把类型签名相关挪到TypeUtils，并开放给编辑器使用
